Revert "rename: shared 폴더로 types 이동"
This reverts commit bdd039e2ee6bb68cd25af364f647b28d6167d42c.
